refactor: move load_encrypted_keyring function under keying module
