+++ scripts/lochness_check_config.py
@@ -3,10 +3,7 @@
 import argparse as ap
 from lochness import config
 from lochness import keyring
-import getpass as gp
 import os
-import cryptease as crypt
-import yaml
 import json
 
 
@@ -14,19 +11,6 @@ def check_config(config_loc: str):
     with open(config_loc, 'rb') as fp:
         config_loaded = config._read_config_file(fp)
     keyring.pretty_print_dict(config_loaded)
-
-
-def load_encrypted_keyring(enc_keyring_loc: str) -> dict:
-    with open(enc_keyring_loc, 'rb') as fp:
-        passphrase = gp.getpass('enter passphrase: ')
-        key = crypt.key_from_file(fp, passphrase)
-        content = b''
-        for chunk in crypt.decrypt(fp, key):
-            content += chunk
-
-        keyring_dict = yaml.load(content, Loader=yaml.FullLoader)
-
-    return keyring_dict
 
 
 def check_keyring(keyring_dict):
@@ -56,7 +40,7 @@ def check_lochness_configurations():
         check_config(args.config)
 
     if args.keyring_encrypted:
-        keyring_dict = load_encrypted_keyring(args.keyring_encrypted)
+        keyring_dict = keyring.load_encrypted_keyring(args.keyring_encrypted)
         check_keyring(keyring_dict)
 
     if args.keyring:
